requirements.txt
google-api-python-client==1.5.3
oauth2client==3.0.0
flask==0.11.1
google-cloud==0.22.0
pandas==0.19.2
numpy==1.11.3
scipy==0.18.1
scikit-learn==0.18.1
requests==2.12.4
